rename server mock data folder
